Move binary and unary op node binder code to BinderUtils
